requirements.txt
cerberus
aioeasywebdav==2.4.0
aiohttp==3.6.2
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
blindspin==2.0.1
boto3==1.12.3
botocore==1.15.3
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.3
ConfigArgParse==1.0
connexion==2020.0.dev1
crayons==0.3.0
cryptography==2.8
datrie==0.8
decorator==4.4.1
docker==4.2.0
docutils==0.15.2
dropbox==9.4.0
filechunkio==1.8
Flask==1.1.1
ftputil==3.4
gevent==1.4.0
gitdb2==3.0.2
GitPython==3.0.8
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.4
networkx==2.4
numpy==1.18.1
openapi-spec-validator==0.2.8
packaging==20.1
pandas==1.0.1
paramiko==2.7.1
pathlib2==2.3.5
pluggy==0.13.0
prettytable==0.7.2
protobuf==3.11.4
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pygraphviz==1.5
pylint==2.4.4
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
pysftp==0.2.9
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.1
python-irodsclient==0.8.2
pytz==2019.3
PyYAML==5.3
ratelimiter==1.2.0
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
smmap2==2.0.5
snakemake==5.8.2
SQLAlchemy==1.3.13
swagger-ui-bundle==0.0.6
testcontainers==2.5.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.7
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.0
xmlrunner==1.7.7
yarl==1.3.0
zipp==2.2.0
